fix: split oversized test_test_filter.py and remove double git call in build_test_scope
Move TestCheckBucketAContentAware and TestBuildTestScopeContentAware to
test_test_filter_content_aware.py to keep test_test_filter.py under 1000 lines.
Eliminate redundant _is_only_version_changes_in_diff call in build_test_scope:
check_bucket_a_content_aware already verified version-only changes, so we can
remove version_bump_in_bucket_a from changed_files directly without a second git diff.
